rename: 각 모듈의 책임에 맞게 파일 위치 변경
- 서버를 시작하는 MusicApplication을 music 모듈에서 api 모듈로 이동
- s3 관련 로직을 music 모듈에서 s3 모듈로 이동
- 음원 관련 검증 로직을 com.music.infra.validator 패키지에서 com.music.validator로 이동
